Move submitter code to separate module.
