+++ rs/cli/src/commands/nodes/remove.rs
@@ -3,7 +3,7 @@ use decentralization::subnets::NodesRemover;
 
 use crate::{
     commands::{AuthRequirement, ExecutableCommand},
-    forum::{ForumPostKind, SubmissionParameters, Submitter},
+    forum::ForumPostKind, submitter::{SubmissionParameters, Submitter},
 };
 
 #[derive(Args, Debug)]
