+++ rs/cli/src/commands/update_authorized_subnets.rs
@@ -10,8 +10,9 @@ use itertools::Itertools;
 use log::info;
 
 use crate::{
-    forum::{ForumPostKind, SubmissionParameters, Submitter},
+    forum::ForumPostKind,
     ic_admin::{IcAdminProposal, IcAdminProposalCommand, IcAdminProposalOptions},
+    submitter::{SubmissionParameters, Submitter},
 };
 
 use super::{AuthRequirement, ExecutableCommand};
