+++ rs/cli/src/commands/version/revise/host_os.rs
@@ -2,7 +2,7 @@ use clap::{error::ErrorKind, Args};
 
 use crate::{
     commands::{AuthRequirement, ExecutableCommand},
-    forum::{ForumPostKind, SubmissionParameters, Submitter},
+    forum::ForumPostKind, submitter::{SubmissionParameters, Submitter},
 };
 
 #[derive(Debug, Args)]
