+++ rs/cli/src/unit_tests/update_unassigned_nodes.rs
@@ -5,8 +5,9 @@ use crate::auth::Neuron;
 use crate::commands::{update_unassigned_nodes::UpdateUnassignedNodes, ExecutableCommand};
 use crate::confirm::ConfirmationModeOptions;
 use crate::cordoned_feature_fetcher::MockCordonedFeatureFetcher;
-use crate::forum::{ForumParameters, SubmissionParameters};
+use crate::forum::ForumParameters;
 use crate::ic_admin::MockIcAdmin;
+use crate::submitter::SubmissionParameters;
 use ic_management_backend::health::MockHealthStatusQuerier;
 use ic_management_backend::{lazy_git::MockLazyGit, lazy_registry::MockLazyRegistry, proposal::MockProposalAgent};
 use ic_management_types::{Network, Subnet};
